requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
pandas==2.2.2
statsmodels==0.14.4
arch==7.2.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.0
networkx==3.2.1
db-dtypes==1.5.0
